# server/api/v1/app/warehouse.go
package app

import (
	
	"github.com/flipped-aurora/gin-vue-admin/server/global"
    "github.com/flipped-aurora/gin-vue-admin/server/model/common/response"
    "github.com/flipped-aurora/gin-vue-admin/server/model/app"
    appReq "github.com/flipped-aurora/gin-vue-admin/server/model/app/request"
    "github.com/gin-gonic/gin"
    "go.uber.org/zap"
)

type WarehouseApi struct {}



// CreateWarehouse 创建仓库
// @Tags Warehouse
// @Summary 创建仓库
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Param data body app.Warehouse true "创建仓库"
// @Success 200 {object} response.Response{msg=string} "创建成功"
// @Router /warehouse/createWarehouse [post]
func (warehouseApi *WarehouseApi) CreateWarehouse(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

	var warehouse app.Warehouse
	err := c.ShouldBindJSON(&warehouse)
	if err != nil {
		response.FailWithMessage(err.Error(), c)
		return
	}
	err = warehouseService.CreateWarehouse(ctx,&warehouse)
	if err != nil {
        global.GVA_LOG.Error("创建失败!", zap.Error(err))
		response.FailWithMessage("创建失败:" + err.Error(), c)
		return
	}
    response.OkWithMessage("创建成功", c)
}

// DeleteWarehouse 删除仓库
// @Tags Warehouse
// @Summary 删除仓库
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Param data body app.Warehouse true "删除仓库"
// @Success 200 {object} response.Response{msg=string} "删除成功"
// @Router /warehouse/deleteWarehouse [delete]
func (warehouseApi *WarehouseApi) DeleteWarehouse(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

	ID := c.Query("ID")
	err := warehouseService.DeleteWarehouse(ctx,ID)
	if err != nil {
        global.GVA_LOG.Error("删除失败!", zap.Error(err))
		response.FailWithMessage("删除失败:" + err.Error(), c)
		return
	}
	response.OkWithMessage("删除成功", c)
}

// DeleteWarehouseByIds 批量删除仓库
// @Tags Warehouse
// @Summary 批量删除仓库
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Success 200 {object} response.Response{msg=string} "批量删除成功"
// @Router /warehouse/deleteWarehouseByIds [delete]
func (warehouseApi *WarehouseApi) DeleteWarehouseByIds(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

	IDs := c.QueryArray("IDs[]")
	err := warehouseService.DeleteWarehouseByIds(ctx,IDs)
	if err != nil {
        global.GVA_LOG.Error("批量删除失败!", zap.Error(err))
		response.FailWithMessage("批量删除失败:" + err.Error(), c)
		return
	}
	response.OkWithMessage("批量删除成功", c)
}

// UpdateWarehouse 更新仓库
// @Tags Warehouse
// @Summary 更新仓库
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Param data body app.Warehouse true "更新仓库"
// @Success 200 {object} response.Response{msg=string} "更新成功"
// @Router /warehouse/updateWarehouse [put]
func (warehouseApi *WarehouseApi) UpdateWarehouse(c *gin.Context) {
    // 从ctx获取标准context进行业务行为
    ctx := c.Request.Context()

	var warehouse app.Warehouse
	err := c.ShouldBindJSON(&warehouse)
	if err != nil {
		response.FailWithMessage(err.Error(), c)
		return
	}
	err = warehouseService.UpdateWarehouse(ctx,warehouse)
	if err != nil {
        global.GVA_LOG.Error("更新失败!", zap.Error(err))
		response.FailWithMessage("更新失败:" + err.Error(), c)
		return
	}
	response.OkWithMessage("更新成功", c)
}

// FindWarehouse 用id查询仓库
// @Tags Warehouse
// @Summary 用id查询仓库
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Param ID query uint true "用id查询仓库"
// @Success 200 {object} response.Response{data=app.Warehouse,msg=string} "查询成功"
// @Router /warehouse/findWarehouse [get]
func (warehouseApi *WarehouseApi) FindWarehouse(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

	ID := c.Query("ID")
	rewarehouse, err := warehouseService.GetWarehouse(ctx,ID)
	if err != nil {
        global.GVA_LOG.Error("查询失败!", zap.Error(err))
		response.FailWithMessage("查询失败:" + err.Error(), c)
		return
	}
	response.OkWithData(rewarehouse, c)
}
// GetWarehouseList 分页获取仓库列表
// @Tags Warehouse
// @Summary 分页获取仓库列表
// @Security ApiKeyAuth
// @Accept application/json
// @Produce application/json
// @Param data query appReq.WarehouseSearch true "分页获取仓库列表"
// @Success 200 {object} response.Response{data=response.PageResult,msg=string} "获取成功"
// @Router /warehouse/getWarehouseList [get]
func (warehouseApi *WarehouseApi) GetWarehouseList(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

	var pageInfo appReq.WarehouseSearch
	err := c.ShouldBindQuery(&pageInfo)
	if err != nil {
		response.FailWithMessage(err.Error(), c)
		return
	}
	list, total, err := warehouseService.GetWarehouseInfoList(ctx,pageInfo)
	if err != nil {
	    global.GVA_LOG.Error("获取失败!", zap.Error(err))
        response.FailWithMessage("获取失败:" + err.Error(), c)
        return
    }
    response.OkWithDetailed(response.PageResult{
        List:     list,
        Total:    total,
        Page:     pageInfo.Page,
        PageSize: pageInfo.PageSize,
    }, "获取成功", c)
}

// GetWarehousePublic 不需要鉴权的仓库接口
// @Tags Warehouse
// @Summary 不需要鉴权的仓库接口
// @Accept application/json
// @Produce application/json
// @Success 200 {object} response.Response{data=object,msg=string} "获取成功"
// @Router /warehouse/getWarehousePublic [get]
func (warehouseApi *WarehouseApi) GetWarehousePublic(c *gin.Context) {
    // 创建业务用Context
    ctx := c.Request.Context()

    // 此接口不需要鉴权
    // 示例为返回了一个固定的消息接口，一般本接口用于C端服务，需要自己实现业务逻辑
    warehouseService.GetWarehousePublic(ctx)
    response.OkWithDetailed(gin.H{
       "info": "不需要鉴权的仓库接口信息",
    }, "获取成功", c)
}