# server/docs/swagger.yaml
definitions:
...
  app.Purchases:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      productID:
        description: productID
        type: string
      purchaseID:
        description: purchaseID
        type: string
      quantity:
        description: quantity
        type: integer
      status:
        description: status
        type: integer
      supplierID:
        description: supplierID
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userID:
        description: userID
        type: integer
      userName:
        description: userName
        type: string
      warehouseID:
        description: warehouseID
        type: string
      warehouseName:
        description: warehouseName
        type: string
    type: object
...
  app.Warehouse:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      name:
        description: name
        type: string
      status:
        description: status
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userID:
        description: userID
        type: integer
      warehouseID:
        description: warehouseID
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.0
paths:
...
  /purchases/createPurchases:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建purchases
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Purchases'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建purchases
      tags:
      - Purchases
  /purchases/deletePurchases:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除purchases
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Purchases'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除purchases
      tags:
      - Purchases
  /purchases/deletePurchasesByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除purchases
      tags:
      - Purchases
  /purchases/findPurchases:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询purchases
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.Purchases'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询purchases
      tags:
      - Purchases
  /purchases/getPurchasesList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: productID
        type: string
      - in: query
        name: purchaseID
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: status
        type: integer
      - in: query
        name: supplierID
        type: string
      - in: query
        name: userID
        type: integer
      - in: query
        name: userName
        type: string
      - in: query
        name: warehouseID
        type: string
      - in: query
        name: warehouseName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取purchases列表
      tags:
      - Purchases
...
  /purchases/updatePurchases:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新purchases
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Purchases'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新purchases
      tags:
      - Purchases
...
  /warehouse/createWarehouse:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建仓库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Warehouse'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建仓库
      tags:
      - Warehouse
  /warehouse/deleteWarehouse:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除仓库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Warehouse'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除仓库
      tags:
      - Warehouse
  /warehouse/deleteWarehouseByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除仓库
      tags:
      - Warehouse
  /warehouse/findWarehouse:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询仓库
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.Warehouse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询仓库
      tags:
      - Warehouse
  /warehouse/getWarehouseList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: userID
        type: integer
      - in: query
        name: warehouseID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取仓库列表
      tags:
      - Warehouse
...
  /warehouse/updateWarehouse:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新仓库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Warehouse'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新仓库
      tags:
      - Warehouse
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...